+++ elastic-job-lite/elastic-job-lite-spring/src/main/java/com/dangdang/ddframe/job/lite/spring/namespace/constants/BaseJobBeanDefinitionParserTag.java
@@ -70,5 +70,5 @@ public final class BaseJobBeanDefinitionParserTag {
     
     public static final String JOB_EXCEPTION_HANDLER_ATTRIBUTE = "job-exception-handler";
     
-    public static final String EVENT_TRACE_DATA_SOURCE_ATTRIBUTE = "event-trace-data-source";
+    public static final String EVENT_TRACE_RDB_DATA_SOURCE_ATTRIBUTE = "event-trace-rdb-data-source";
 }